requirements.txt
praw==7.1.4
matplotlib==3.3.2
seaborn==0.11.0
numpy==1.19.2
streamlit==0.77.0
pandas==1.1.3
Pillow==8.1.2
datetime==4.3.0
altair_viewer==0.3.0